homework2/requirements.txt
streamlit
Pillow
gtts